core/rawdb: move buffer code to batch file
